LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-history-interface/005-expected.txt


PASS history.pushState support is needed for this testcase
PASS <body onpopstate="..."> should register a listener for the popstate event
PASS window.onpopstate should register a listener for the popstate event
